Module checksums::ops

Main functions doing actual work.

Use create_hashes() to prepare the hashes for a path.

Then use write_hashes() to save it to disk, or read_hashes() to get the saved hashes and compare them.

Functions

create_hashes
Create subpath->hash mappings for a given path using a given algorithm up to a given depth.

write_hashes
Serialise the specified hashes to the specified output file.